Refactor wrapper
